+++ lib/l10n/app_localizations_en.dart
@@ -740,6 +740,9 @@ class AppLocalizationsEn extends AppLocalizations {
   String get configLabelTopic => 'Topic';
 
   @override
+  String get configLabelDirectMessage => 'Direct message';
+
+  @override
   String get exportImportTitle => 'Export / Import';
 
   @override