+++ lib/l10n/app_localizations_eo.dart
@@ -742,6 +742,9 @@ class AppLocalizationsEo extends AppLocalizations {
   String get configLabelTopic => 'Topic';
 
   @override
+  String get configLabelDirectMessage => 'Direct message';
+
+  @override
   String get exportImportTitle => 'Export / Import';
 
   @override