+++ lib/l10n/app_localizations_tlh.dart
@@ -729,6 +729,9 @@ class AppLocalizationsTlh extends AppLocalizations {
   String get configLabelTopic => 'Topic';
 
   @override
+  String get configLabelDirectMessage => 'Direct message';
+
+  @override
   String get exportImportTitle => 'Export / Import';
 
   @override