+++ lib/l10n/app_localizations_tr.dart
@@ -745,6 +745,9 @@ class AppLocalizationsTr extends AppLocalizations {
   String get configLabelTopic => 'Topic';
 
   @override
+  String get configLabelDirectMessage => 'Direct message';
+
+  @override
   String get exportImportTitle => 'Export / Import';
 
   @override